+++ src/main/java/chess/domain/chess/ChessGame.java
@@ -1,8 +1,8 @@
 package chess.domain.chess;
 
 import chess.domain.board.ChessBoard;
+import chess.domain.piece.Piece;
 import chess.domain.piece.Position;
-import chess.domain.piece.type.Piece;
 
 import java.util.Map;
