+++ src/test/java/chess/domain/board/ChessBoardTest.java
@@ -2,11 +2,11 @@ package chess.domain.board;
 
 import chess.domain.chess.CampType;
 import chess.domain.chess.ChessGame;
+import chess.domain.piece.Piece;
+import chess.domain.piece.PieceType;
 import chess.domain.piece.Position;
 import chess.domain.piece.move.piece.PawnMove;
 import chess.domain.piece.move.piece.RookMove;
-import chess.domain.piece.type.Piece;
-import chess.domain.piece.type.PieceType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
